update host: use looser matching

+++ ailib/common/__init__.py
@@ -158,13 +158,16 @@ def create_onprem(overrides={}, debug=False):
         ip = overrides.get('ip') or get_ip() or '192.168.122.1'
         ipv6 = ':' in ip
         info(f"Using ip {ip}")
-        with open(f"{tmpdir}/pod.yml", 'w') as p:
-            pod_url = "https://raw.githubusercontent.com/openshift/assisted-service/master/deploy/podman/pod.yml"
-            response = urllib.request.urlopen(pod_url).read().decode('utf-8')
-            response = response.replace('latest', onprem_version).replace(f'centos7:{onprem_version}', 'centos7:latest')
-            if ipv6:
-                response = response.replace('127.0.0.1', f"[{ip}")
-            p.write(response)
+        if os.path.exists('pod.yml'):
+            info("Using existing pod.yml")
+            copy2("pod.yml", tmpdir)
+        else:
+            with open(f"{tmpdir}/pod.yml", 'w') as p:
+                pod_url = "https://raw.githubusercontent.com/openshift/assisted-service/master/deploy/podman/pod.yml"
+                response = urllib.request.urlopen(pod_url).read().decode('utf-8')
+                response = response.replace('latest', onprem_version).replace(f'centos7:{onprem_version}',
+                                                                              'centos7:latest')
+                p.write(response)
         if os.path.exists('configmap.yml'):
             info("Using existing configmap.yml")
             copy2("configmap.yml", tmpdir)
@@ -173,8 +176,10 @@ def create_onprem(overrides={}, debug=False):
                 cm_name = 'okd-configmap' if overrides.get('okd', False) else 'configmap'
                 cm_url = "https://raw.githubusercontent.com/openshift/assisted-service/master/deploy/podman/"
                 cm_url += f"{cm_name}.yml"
-                response = urllib.request.urlopen(cm_url)
-                c.write(response.read().decode('utf-8'))
+                response = urllib.request.urlopen(cm_url).read().decode('utf-8')
+                if ipv6:
+                    response = response.replace('127.0.0.1:8090', f'"[{ip}]:8090"')
+                c.write(response)
             if ipv6 and '[' not in ip:
                 ip = f"[{ip}]"
             IMAGE_SERVICE_BASE_URL = f'http://{ip}:8888'
@@ -198,7 +203,7 @@ def create_onprem(overrides={}, debug=False):
             call(cmd, shell=True)
         storage = '--storage-driver vfs' if 'KUBERNETES_SERVICE_PORT' in os.environ else ''
         network = '--network assistedv6' if ipv6 else ''
-        cmd = f"podman {storage} {network} play kube --replace --configmap {tmpdir}/configmap.yml {tmpdir}/pod.yml"
+        cmd = f"podman {storage} play kube {network} --replace --configmap {tmpdir}/configmap.yml {tmpdir}/pod.yml"
         info(f"Running: {cmd}")
         call(cmd, shell=True)
